re-do migration, use smallint

+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Commands/Create/CreateDialogCommand.cs
@@ -116,12 +116,12 @@ internal sealed class CreateDialogCommandHandler : IRequestHandler<CreateDialogC
             maxDepth: 20,
             maxWidth: 20));
 
-        dialog.FromPartyTransmissionsCount = dialog.Transmissions
+        dialog.FromPartyTransmissionsCount = (short)dialog.Transmissions
             .Count(x => x.TypeId
                 is DialogTransmissionType.Values.Submission
                 or DialogTransmissionType.Values.Correction);
 
-        dialog.FromServiceOwnerTransmissionsCount = dialog.Transmissions
+        dialog.FromServiceOwnerTransmissionsCount = (short)dialog.Transmissions
             .Count(x => x.TypeId is not
                 (DialogTransmissionType.Values.Submission
                 or DialogTransmissionType.Values.Correction));
